pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = "REST API for managing contacts"
authors = ["Dmytro Zhuk <dm.zhuk@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9 || ^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.9"
pydantic = { version = "^1.9.0", extras = ["email"] }
python-dateutil = "^2.9.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-dotenv = "^1.0.1"
aiosmtplib = "^3.0.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 